# sms_api/api/spec/openapi_3_1_0_generated.yaml
openapi: 3.1.0
info:
  title: sms-api
  version: 0.2.20-dev
paths:
...
  /experiments/launch:
    post:
      tags:
      - Simulations - vEcoli
      summary: Launches a nextflow-powered vEcoli simulation workflow
      operationId: run-experiment
      parameters:
      - name: config_id
        in: query
        required: false
        schema:
          type: string
          description: Configuration ID of an existing available vecoli simulation
            configuration JSON
          default: sms
          title: Config Id
        description: Configuration ID of an existing available vecoli simulation configuration
          JSON
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_run-experiment'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EcoliExperimentDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /experiments/get:
    get:
      tags:
      - Simulations - vEcoli
      summary: Get Experiment
      operationId: get-experiment
      parameters:
      - name: experiment_id
        in: query
        required: true
        schema:
          type: string
          title: Experiment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EcoliExperimentDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /experiments/remove:
    delete:
      tags:
      - Simulations - vEcoli
      summary: Delete Experiment
      operationId: delete-experiment
      parameters:
      - name: experiment_id
        in: query
        required: true
        schema:
          type: string
          title: Experiment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: string
                title: Response Delete-Experiment
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /experiments/status:
    get:
      tags:
      - Simulations - vEcoli
      summary: Get the simulation status record by its ID
      operationId: get-vecoli-simulation-status
      parameters:
      - name: experiment_tag
        in: query
        required: true
        schema:
          type: string
          title: Experiment Tag
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationRun'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /experiments/log:
    post:
      tags:
      - Simulations - vEcoli
      summary: Get the simulation log record of a given experiment
      operationId: get-vecoli-simulation-log
      parameters:
      - name: experiment_id
        in: query
        required: true
        schema:
          type: string
          title: Experiment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: string
                title: Response Get-Vecoli-Simulation-Log
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /experiments/data:
    post:
      tags:
      - Simulations - vEcoli
      summary: Get simulation outputs
      operationId: get-data
      parameters:
      - name: experiment_id
        in: query
        required: true
        schema:
          type: string
          description: Experiment id for the simulation.
          title: Experiment Id
        description: Experiment id for the simulation.
      - name: variant_id
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Variant Id
      - name: lineage_seed_id
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Lineage Seed Id
      - name: generation_id
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Generation Id
      - name: agent_id
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Agent Id
      - name: filename
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Name you wish to assign to the downloaded zip file
          title: Filename
        description: Name you wish to assign to the downloaded zip file
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_run-experiment:
      properties:
        overrides:
          anyOf:
          - $ref: '#/components/schemas/ConfigOverrides'
          - type: 'null'
        metadata:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Metadata
      type: object
      title: Body_run-experiment
...
    ConfigOverrides:
      properties:
        experiment_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Experiment Id
        sim_data_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Sim Data Path
        suffix_time:
          type: boolean
          title: Suffix Time
          default: false
        parca_options:
          additionalProperties:
            anyOf:
            - type: boolean
            - type: integer
            - type: string
            - {}
            - type: 'null'
          type: object
          title: Parca Options
        generations:
          type: integer
          title: Generations
          default: 1
        n_init_sims:
          anyOf:
          - type: integer
          - type: 'null'
          title: N Init Sims
        max_duration:
          type: number
          title: Max Duration
          default: 10800.0
        initial_global_time:
          type: number
          title: Initial Global Time
          default: 0.0
        time_step:
          type: number
          title: Time Step
          default: 1.0
        single_daughters:
          type: boolean
          title: Single Daughters
          default: true
        emitter:
          type: string
          title: Emitter
          default: parquet
        emitter_arg:
          additionalProperties:
            type: string
          type: object
          title: Emitter Arg
        variants:
          additionalProperties:
            additionalProperties:
              additionalProperties:
                items:
                  anyOf:
                  - type: number
                  - type: string
                  - type: integer
                type: array
              type: object
            type: object
          type: object
          title: Variants
        analysis_options:
          additionalProperties: true
          type: object
          title: Analysis Options
        gcloud:
          anyOf:
          - type: string
          - type: 'null'
          title: Gcloud
        agent_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Agent Id
        parallel:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Parallel
        divide:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Divide
        d_period:
          anyOf:
          - type: boolean
          - type: 'null'
          title: D Period
        division_threshold:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Division Threshold
        division_variable:
          items:
            type: string
          type: array
          title: Division Variable
        chromosome_path:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Chromosome Path
        spatial_environment:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Spatial Environment
        fixed_media:
          anyOf:
          - type: string
          - type: 'null'
          title: Fixed Media
        condition:
          anyOf:
          - type: string
          - type: 'null'
          title: Condition
        save:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Save
        save_times:
          items:
            type: string
          type: array
          title: Save Times
        add_processes:
          items:
            type: string
          type: array
          title: Add Processes
        exclude_processes:
          items:
            type: string
          type: array
          title: Exclude Processes
        profile:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Profile
        processes:
          items:
            type: string
          type: array
          title: Processes
        process_configs:
          additionalProperties: true
          type: object
          title: Process Configs
        topology:
          additionalProperties: true
          type: object
          title: Topology
        engine_process_reports:
          items:
            items:
              type: string
            type: array
          type: array
          title: Engine Process Reports
        emit_paths:
          items:
            type: string
          type: array
          title: Emit Paths
        progress_bar:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Progress Bar
        emit_topology:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Emit Topology
        emit_processes:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Emit Processes
        emit_config:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Emit Config
        emit_unique:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Emit Unique
        log_updates:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Log Updates
        raw_output:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Raw Output
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        seed:
          anyOf:
          - type: integer
          - type: 'null'
          title: Seed
        mar_regulon:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Mar Regulon
        amp_lysis:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Amp Lysis
        initial_state_file:
          anyOf:
          - type: string
          - type: 'null'
          title: Initial State File
        skip_baseline:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Skip Baseline
        daughter_outdir:
          anyOf:
          - type: string
          - type: 'null'
          title: Daughter Outdir
        lineage_seed:
          anyOf:
          - type: integer
          - type: 'null'
          title: Lineage Seed
        fail_at_max_duration:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Fail At Max Duration
        inherit_from:
          items:
            type: string
          type: array
          title: Inherit From
        spatial_environment_config:
          additionalProperties: true
          type: object
          title: Spatial Environment Config
        swap_processes:
          additionalProperties: true
          type: object
          title: Swap Processes
        flow:
          additionalProperties: true
          type: object
          title: Flow
        initial_state_overrides:
          items:
            type: string
          type: array
          title: Initial State Overrides
        initial_state:
          additionalProperties: true
          type: object
          title: Initial State
      type: object
      title: ConfigOverrides
    EcoliExperimentDTO:
      properties:
        experiment_id:
          type: string
          title: Experiment Id
        request:
          $ref: '#/components/schemas/EcoliExperimentRequestDTO'
        last_updated:
          type: string
          title: Last Updated
        metadata:
          additionalProperties:
            type: string
          type: object
          title: Metadata
        experiment_tag:
          anyOf:
          - type: string
          - type: 'null'
          title: Experiment Tag
      type: object
      required:
      - experiment_id
      - request
      title: EcoliExperimentDTO
    EcoliExperimentRequestDTO:
      properties:
        config_id:
          type: string
          title: Config Id
        overrides:
          anyOf:
          - $ref: '#/components/schemas/ConfigOverrides'
          - type: 'null'
      type: object
      required:
      - config_id
      title: EcoliExperimentRequestDTO
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    JobStatus:
      type: string
      enum:
      - waiting
      - queued
      - running
      - completed
      - failed
      title: JobStatus
...
    SimulationRun:
      properties:
        id:
          type: string
          title: Id
        status:
          $ref: '#/components/schemas/JobStatus'
      type: object
      required:
      - id
      - status
      title: SimulationRun
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError